+++ FortnoxAPILibrary/Connectors/ContractAccrualConnector.cs
@@ -26,7 +26,7 @@ namespace FortnoxAPILibrary.Connectors
 		/// </summary>
 		/// <param name="id">Identifier of the contractAccrual to find</param>
 		/// <returns>The found contractAccrual</returns>
-		public ContractAccrual Get(int? id)
+		public ContractAccrual Get(long? id)
 		{
 			return GetAsync(id).Result;
 		}
@@ -55,7 +55,7 @@ namespace FortnoxAPILibrary.Connectors
 		/// Deletes a contractAccrual
 		/// </summary>
 		/// <param name="id">Identifier of the contractAccrual to delete</param>
-		public void Delete(int? id)
+		public void Delete(long? id)
 		{
 			DeleteAsync(id).Wait();
 		}
@@ -73,7 +73,7 @@ namespace FortnoxAPILibrary.Connectors
 		{
 			return await BaseFind().ConfigureAwait(false);
 		}
-		public async Task DeleteAsync(int? id)
+		public async Task DeleteAsync(long? id)
 		{
 			await BaseDelete(id.ToString()).ConfigureAwait(false);
 		}
@@ -85,7 +85,7 @@ namespace FortnoxAPILibrary.Connectors
 		{
 			return await BaseUpdate(contractAccrual, contractAccrual.DocumentNumber.ToString()).ConfigureAwait(false);
 		}
-		public async Task<ContractAccrual> GetAsync(int? id)
+		public async Task<ContractAccrual> GetAsync(long? id)
 		{
 			return await BaseGet(id.ToString()).ConfigureAwait(false);
 		}
