+++ FortnoxAPILibrary/Connectors/LabelConnector.cs
@@ -46,7 +46,7 @@ namespace FortnoxAPILibrary.Connectors
 		/// Deletes a label
 		/// </summary>
 		/// <param name="id">Identifier of the label to delete</param>
-		public void Delete(int? id)
+		public void Delete(long? id)
 		{
 			DeleteAsync(id).Wait();
 		}
@@ -64,7 +64,7 @@ namespace FortnoxAPILibrary.Connectors
 		{
 			return await BaseFind().ConfigureAwait(false);
 		}
-		public async Task DeleteAsync(int? id)
+		public async Task DeleteAsync(long? id)
 		{
 			await BaseDelete(id.ToString()).ConfigureAwait(false);
 		}
